Added long-reset command to set both sleep and wake
Added long-reset command to set both sleep and wake, since that's
what most people will want to do. Also added a bash script that
makes it simple to do an 8 second reset with a 2 second delay.
This script talks to the power controller over i2c and calls the
unlock command if your kernel major version is >= 4.